Filter out results instead of throwing an error
Currently the server returns an empty list/error if you do not have the access rights to query one of the triples in the results. Instead of rejecting the whole triple we could use a code snippet like this to filter out all the results the issuer does not have access to.